+++ ui/client/reducers/nodeDetailsState.ts
@@ -1,7 +1,12 @@
 import {Action} from "../actions/reduxTypes"
 import {ValidationData} from "../actions/nk"
+import {NodeValidationError, UIParameter} from "../types"
 
-export type NodeDetailsState = ValidationData
+export type NodeDetailsState = {
+    parameters? : Map<string, UIParameter>,
+    validationErrors: NodeValidationError[],
+    validationPerformed: boolean,
+}
 
 const initialState: NodeDetailsState = {
   validationErrors: [],