+++ ui/client/types/index.ts
@@ -3,4 +3,5 @@ export * from "./groups"
 export * from "./edge"
 export * from "./process"
 export * from "./validation"
+export * from "./definition"