+++ ui/server/src/main/scala/pl/touk/nussknacker/ui/api/NodesResources.scala
@@ -10,7 +10,7 @@ import pl.touk.nussknacker.engine.ModelData
 import pl.touk.nussknacker.engine.ProcessingTypeData.ProcessingType
 import pl.touk.nussknacker.engine.additionalInfo.{NodeAdditionalInfo, NodeAdditionalInfoProvider}
 import pl.touk.nussknacker.engine.api.MetaData
-import pl.touk.nussknacker.engine.api.context.ValidationContext
+import pl.touk.nussknacker.engine.api.context.{ProcessCompilationError, ValidationContext}
 import pl.touk.nussknacker.engine.api.process.ParameterConfig
 import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
 import pl.touk.nussknacker.engine.compile.nodevalidation.{NodeDataValidator, ValidationNotPerformed, ValidationPerformed}
@@ -25,6 +25,7 @@ import pl.touk.nussknacker.restmodel.displayedgraph.ProcessProperties
 import pl.touk.nussknacker.ui.validation.PrettyValidationErrors
 import io.circe.generic.semiauto.deriveDecoder
 import org.springframework.util.ClassUtils
+import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.MissingParameters
 import pl.touk.nussknacker.engine.api.typed.TypingResultDecoder
 import pl.touk.nussknacker.ui.definition.UIParameter
 
@@ -68,7 +69,13 @@ class NodesResources(val processRepository: FetchingProcessRepository[Future],
                 case ValidationNotPerformed => NodeValidationResult(None, Nil, validationPerformed = false)
                 case ValidationPerformed(errors, parameters) =>
                   val uiParams = parameters.map(_.map(UIParameter(_, ParameterConfig.empty)))
-                  val uiErrors = errors.map(PrettyValidationErrors.formatErrorMessage)
+                  //We don't return MissingParameter error when we are returning those missing parameters to be added - since
+                  //it's not really exception ATM
+                  def shouldIgnoreError(pce: ProcessCompilationError): Boolean = pce match {
+                    case MissingParameters(params, _) => params.forall(missing => uiParams.exists(_.exists(_.name == missing)))
+                    case _ => false
+                  }
+                  val uiErrors = errors.filterNot(shouldIgnoreError).map(PrettyValidationErrors.formatErrorMessage)
                   NodeValidationResult(uiParams, uiErrors, validationPerformed = true)
               }
             }